[QPE-223] Initial build pipeline commit
