+++ .circleci/config.yml
@@ -54,3 +54,4 @@ workflows:
             branches:
               only:
                 - master
+                - feature/QPE-223
