--- extension/br.com.clever.wordcloud.qext
@@ -1,9 +1,0 @@
-{
-	"name": "Word Cloud",
-	"description": "Word Cloud",
-	"icon": "treemap",
-	"type": "visualization",
-	"version": 1,
-	"preview": "wordcloud.png",
-	"author": "Clever Anjos (clever@clever.com.br)"
-}
