+++ package.json
@@ -1,0 +1,49 @@
+{
+  "name": "qlik-word-cloud",
+  "version": "0.0.1",
+  "description": "Qlik Sense Extension Word Cloud",
+  "keywords": "word cloud qliksense extension",
+  "repository": "https://github.com/cleveranjos/br.com.clever.wordcloud",
+  "author": "Clever Anjos <clever@clever.com.br>",
+  "license": "Apache License 2.0",
+  "scripts": {
+    "build": "gulp build",
+    "eslint": "eslint src",
+    "eslint:fix": "eslint --fix src",
+    "test": "karma start karma.conf.js",
+    "test-once": "karma start karma.conf.js --single-run",
+    "watch": "gulp watch"
+  },
+  "devDependencies": {
+    "@babel/core": "7.1.2",
+    "@babel/polyfill": "7.0.0",
+    "@babel/preset-env": "7.1.0",
+    "babel-eslint": "10.0.1",
+    "babel-loader": "8.0.4",
+    "copy-webpack-plugin": "4.5.3",
+    "css-loader": "1.0.0",
+    "del": "2.0.2",
+    "eslint": "5.7.0",
+    "eslint-loader": "2.1.1",
+    "gulp": "4.0.0",
+    "gulp-zip": "3.0.2",
+    "jasmine-core": "3.2.1",
+    "jquery": "3.3.1",
+    "karma": "3.0.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-jasmine": "1.1.2",
+    "karma-sourcemap-loader": "0.3.7",
+    "karma-webpack": "3.0.5",
+    "less": "3.8.1",
+    "less-loader": "4.1.0",
+    "style-loader": "0.23.1",
+    "stylelint": "8.4.0",
+    "stylelint-webpack-plugin": "0.10.5",
+    "webpack": "4.20.2",
+    "webpack-cli": "3.1.2",
+    "webpack-dev-server": "3.1.10"
+  },
+  "dependencies": {
+    "d3": "3.5.12"
+  }
+}
